Rewrite auto-sizing code
We try to leverage WxPython functionality as much as possible for the size of the list control. Luckily macOS reduces large windows automatically if needed, so we don't need to worry about that as long as the MinSize is set to a small amount. The window now has a large title like the others.